Flytter tokenprovider og getEnvVar til felles
