+++ aareg/src/main/kotlin/no/nav/helsearbeidsgiver/inntektsmelding/aareg/Environment.kt
@@ -1,5 +1,7 @@
 package no.nav.helsearbeidsgiver.inntektsmelding.aareg
 
+import no.nav.helsearbeidsgiver.felles.getEnvVar
+
 fun setUpEnvironment(): Environment {
     return Environment(
         raw = System.getenv(),
@@ -14,6 +16,3 @@ data class Environment(
     val raw: Map<String, String>,
     val aaregUrl: String
 )
-
-fun getEnvVar(varName: String, defaultValue: String? = null) =
-    System.getenv(varName) ?: defaultValue ?: throw RuntimeException("Environment: Missing required variable \"$varName\"")
