+++ felles/src/main/kotlin/no/nav/helsearbeidsgiver/felles/Utils.kt
@@ -1,0 +1,6 @@
+package no.nav.helsearbeidsgiver.felles
+
+fun getEnvVar(varName: String, defaultValue: String? = null) =
+    System.getenv(varName) ?: defaultValue ?: throw RuntimeException("Environment: Missing required variable \"$varName\"")
+
+fun isLocal(): Boolean = getEnvVar("KTOR_ENV", "local") == "local"
