+++ felles/src/main/kotlin/no/nav/helsearbeidsgiver/felles/oauth2/AzureOAuth2Environment.kt
@@ -1,0 +1,10 @@
+package no.nav.helsearbeidsgiver.felles.oauth2
+
+data class AzureOAuth2Environment(
+    val scope: String,
+    val azureWellKnownUrl: String,
+    val azureTokenEndpointUrl: String,
+    val azureAppClientID: String,
+    val azureAppClientSecret: String,
+    val azureAppJwk: String
+)
