+++ gradle.properties
@@ -13,3 +13,5 @@ ktorVersion=2.1.2
 lettuceVersion=6.2.0.RELEASE
 mockkVersion=1.13.2
 rapidsAndRiversVersion=2022092314391663936769.9d5d33074875
+tokenClientVersion=2.1.4
+tokenproviderVersion=0.2.6
