+++ joark/src/main/kotlin/no/nav/helsearbeidsgiver/inntektsmelding/joark/Environment.kt
@@ -1,5 +1,7 @@
 package no.nav.helsearbeidsgiver.inntektsmelding.joark
 
+import no.nav.helsearbeidsgiver.felles.getEnvVar
+
 fun setUpEnvironment(): Environment {
     return Environment(
         raw = System.getenv(),
@@ -15,8 +17,3 @@ data class Environment(
 data class Dokarkiv(
     val url: String
 )
-
-fun isLocal(): Boolean = getEnvVar("KTOR_ENV", "local") == "local"
-
-fun getEnvVar(varName: String, defaultValue: String? = null) =
-    System.getenv(varName) ?: defaultValue ?: throw RuntimeException("Environment: Missing required variable \"$varName\"")
